+++ grand/__init__.py
@@ -1,16 +1,16 @@
 """
-Grand graph databasifier.
+Grand graphs package.
 
-Aug 2020
 """
 
+from typing import Optional
 from .backends import Backend, NetworkXBackend
 from .dialects import NetworkXDialect, IGraphDialect
 
 
 _DEFAULT_BACKEND = NetworkXBackend
 
-__version__ = "0.4.2"
+__version__ = "0.5.0"
 
 
 class Graph:
@@ -19,32 +19,49 @@ class Graph:
 
     """
 
-    def __init__(
-        self, backend: Backend = None, directed: bool = True, dialects: dict = None
-    ):
+    def __init__(self, backend: Optional[Backend] = None, **backend_kwargs: dict):
         """
         Create a new grand.Graph.
+
+        The only positional argument is the backend to use. All other arguments
+        are passed to the backend's constructor, if a type is provided.
+        Otherwise, kwargs are ignored.
 
         Arguments:
             backend (Backend): The backend to use. If none is provided, will
                 default to _DEFAULT_BACKEND.
 
-        Returns:
-            None
-
         """
-        self.backend = backend or _DEFAULT_BACKEND(directed=directed)
+        self.backend = backend or _DEFAULT_BACKEND
+
+        # If you passed a class instead of an instance, instantiate it with
+        # kwargs from the constructor:
+        if isinstance(self.backend, type):
+            self.backend = self.backend(**backend_kwargs)
 
         # Attach dialects:
         self.nx = NetworkXDialect(self)
         self.igraph = IGraphDialect(self)
 
-        if dialects:
-            self.dialects = {k: v(self) for k, v in dialects.items()}
 
-    def save(self, filename: str) -> str:
-        raise NotImplementedError()
+class DiGraph(Graph):
+    """
+    A grand.DiGraph enables you to manipulate a directed graph. This is a
+    convenience class that inherits from grand.Graph.
 
-    @staticmethod
-    def load(self, filename: str) -> str:
-        raise NotImplementedError()
+    """
+
+    def __init__(self, backend: Optional[Backend] = None, **backend_kwargs: dict):
+        """
+        Create a new grand.DiGraph.
+
+        The only positional argument is the backend to use. All other arguments
+        are passed to the backend's constructor, if a type is provided.
+        Otherwise, kwargs are ignored.
+
+        Arguments:
+            backend (Backend): The backend to use. If none is provided, will
+                default to _DEFAULT_BACKEND.
+
+        """
+        super().__init__(backend, **{**backend_kwargs, "directed": True})