+++ grand/test_graph.py
@@ -1,6 +1,6 @@
 import unittest
 
-from . import Graph, NetworkXBackend
+from . import Graph, DiGraph
 
 
 class TestGraph(unittest.TestCase):
@@ -10,3 +10,7 @@ class TestGraph(unittest.TestCase):
     def test_can_use_nx_backend(self):
         Graph().nx
 
+    def test_can_create_directed(self):
+        assert Graph(directed=True).nx.is_directed() is True
+        assert Graph(directed=False).nx.is_directed() is False
+        assert DiGraph().nx.is_directed() is True